Remove dependency (#209)
* remove algob dependency

* lint:fix

* fix
