+++ examples/permissioned-voting/scripts/vote/vote.js
@@ -1,4 +1,5 @@
-const { TransactionType, SignType, stringToBytes } = require('@algorand-builder/algob');
+const { stringToBytes } = require('@algorand-builder/algob');
+const { TransactionType, SignType } = require('@algorand-builder/runtime/build/types');
 const { executeTransaction } = require('./common');
 
 async function run (runtimeEnv, deployer) {
