+++ examples/permissioned-voting/scripts/voting.js
@@ -1,4 +1,5 @@
-const { executeTransaction, TransactionType, SignType, stringToBytes, uint64ToBigEndian } = require('@algorand-builder/algob');
+const { executeTransaction, stringToBytes, uint64ToBigEndian } = require('@algorand-builder/algob');
+const { TransactionType, SignType } = require('@algorand-builder/runtime/build/types');
 
 async function run (runtimeEnv, deployer) {
   const masterAccount = deployer.accountsByName.get('master-account');
