+++ packages/algob/src/lib/algo-operator.ts
@@ -1,3 +1,9 @@
+import { encodeNote, parseSSCAppArgs } from "@algorand-builder/runtime";
+import type {
+  SSCDeploymentFlags,
+  SSCOptionalFlags,
+  TxParams
+} from "@algorand-builder/runtime/build/types";
 import type { LogicSigArgs } from "algosdk";
 import algosdk from "algosdk";
 
@@ -16,14 +22,10 @@ import type {
   FundASCFlags,
   LsigInfo,
   Network,
-  SSCDeploymentFlags,
   SSCInfo,
-  SSCOptionalFlags,
-  StrMap,
-  TxParams
+  StrMap
 } from "../types";
 import { CompileOp } from "./compile";
-import { parseSSCAppArgs } from "./ssc";
 import * as tx from "./tx";
 const confirmedRound = "confirmed-round";
 
@@ -225,7 +227,7 @@ export class AlgoOperatorImpl implements AlgoOperator {
     console.log(message);
 
     const closeToRemainder = undefined;
-    const note = tx.encodeNote(payFlags.note, payFlags.noteb64);
+    const note = encodeNote(payFlags.note, payFlags.noteb64);
     const t = algosdk.makePaymentTxnWithSuggestedParams(flags.funder.addr, contractAddress,
       flags.fundingMicroAlgo, closeToRemainder,
       note,
