+++ packages/algob/test/stubs/algo-operator.ts
@@ -1,3 +1,4 @@
+import type { SSCDeploymentFlags, SSCOptionalFlags, TxParams } from "@algorand-builder/runtime/build/types";
 import type { LogicSig } from "algosdk";
 import { Algodv2, LogicSigArgs } from "algosdk";
 
@@ -12,10 +13,7 @@ import {
   ASCCache,
   FundASCFlags,
   LsigInfo,
-  SSCDeploymentFlags,
-  SSCInfo,
-  SSCOptionalFlags,
-  TxParams
+  SSCInfo
 } from "../../src/types";
 
 export class AlgoOperatorDryRunImpl implements AlgoOperator {
