+++ packages/runtime/src/errors/errors-list.ts
@@ -13,7 +13,8 @@ export function getErrorCode (error: ErrorDescriptor): string {
 }
 
 export const ERROR_RANGES = {
-  TEAL: { min: 0, max: 99, title: "TEAL opcode errors" }
+  TEAL: { min: 0, max: 99, title: "TEAL opcode errors" },
+  TRANSACTION: { min: 0, max: 99, title: "Transaction error" }
 };
 
 const PARSE_ERROR = "Parse Error";
@@ -248,10 +249,19 @@ by an index that does not exist.`
     description: `lsig validation error`
   },
   INVALID_TRANSACTION_PARAMS: {
-    number: 27,
+    number: 37,
     message: "Secret key and Logic Signature should not be passed together",
     title: "Transaction params error",
     description: `Transaction params error`
+  }
+};
+
+const transactionErrors = {
+  UNKNOWN_TRANSACTION_TYPE: {
+    number: 1,
+    message: "Unknown Transaction type: %transaction%",
+    title: "Transaction error",
+    description: `Transaction error`
   }
 };
 
@@ -260,5 +270,6 @@ export const ERRORS: {
     [errorName: string]: ErrorDescriptor
   };
 } = {
-  TEAL: tealErrors
+  TEAL: tealErrors,
+  TRANSACTION: transactionErrors
 };
