+++ packages/runtime/src/index.ts
@@ -1,9 +1,19 @@
-import { Interpreter } from "../src/interpreter/interpreter";
 import { StoreAccount } from "./account";
+import { applyErrorMessageTemplate } from "./errors/errors";
+import { Interpreter } from "./interpreter/interpreter";
+import { addressToPk, parseSSCAppArgs, stringToBytes, uint64ToBigEndian } from "./lib/parsing";
+import { encodeNote, mkTransaction } from "./lib/txn";
 import { Runtime } from "./runtime";
 
 export {
   Interpreter,
   Runtime,
-  StoreAccount
+  StoreAccount,
+  mkTransaction,
+  applyErrorMessageTemplate,
+  parseSSCAppArgs,
+  addressToPk,
+  uint64ToBigEndian,
+  stringToBytes,
+  encodeNote
 };
