+++ packages/runtime/src/lib/constants.ts
@@ -1,6 +1,5 @@
-// https://docs.microsoft.com/en-us/dotnet/api/system.uint64.maxvalue?view=net-5.0
-export const MAX_UINT64 = BigInt("18446744073709551615");
-export const MIN_UINT64 = BigInt("0");
+export const MIN_UINT64 = 0n;
+export const MAX_UINT64 = 0xFFFFFFFFFFFFFFFFn;
 export const MAX_UINT8 = 255;
 export const MIN_UINT8 = 0;
 export const DEFAULT_STACK_ELEM = BigInt("0");
