+++ packages/runtime/src/lib/txn.ts
@@ -1,11 +1,11 @@
-import { AssetDefEnc } from "algosdk";
+import algosdk, { AssetDefEnc, SuggestedParams, Transaction } from "algosdk";
 
 import { TealError } from "../errors/errors";
 import { ERRORS } from "../errors/errors-list";
 import { Op } from "../interpreter/opcode";
 import { TxFieldDefaults, TxnFields } from "../lib/constants";
-import { stringToBytes } from "../lib/parsing";
-import { StackElem, TxField, Txn, TxnType } from "../types";
+import { parseSSCAppArgs, stringToBytes } from "../lib/parsing";
+import { ExecParams, StackElem, TransactionType, TxField, Txn, TxnType } from "../types";
 
 const assetTxnFields = new Set([
   'ConfigAssetTotal',
@@ -118,4 +118,107 @@ export function txAppArg (txField: TxField, tx: Txn, idx: number, op: Op,
   throw new TealError(ERRORS.TEAL.INVALID_OP_ARG, {
     opcode: "txna or gtxna"
   });
+}
+
+export function encodeNote (note: string | undefined, noteb64: string| undefined): Uint8Array | undefined {
+  if (note === undefined && noteb64 === undefined) { return undefined; }
+  const encoder = new TextEncoder();
+  return noteb64 ? encoder.encode(noteb64) : encoder.encode(note);
+}
+
+/**
+ * Returns unsigned transaction as per ExecParams
+ * ExecParams can be of following types:
+ *  + AlgoTransferParam used for transferring algo
+ *  + AssetTransferParam used for transferring asset
+ *  + SSCCallsParam used for calling stateful smart contracts.
+ For more advanced use-cases, please use `algosdk.tx` directly.
+ NOTE: parseSSCAppArgs is used to handle case when user passes appArgs similar to goal
+ * @param execParams ExecParams
+ * @param suggestedParams Suggested params
+ * @returns SDK Transaction object
+ */
+export function mkTransaction (execParams: ExecParams, suggestedParams: SuggestedParams): Transaction {
+  const note = encodeNote(execParams.payFlags.note, execParams.payFlags.noteb64);
+  const transactionType = execParams.type;
+  switch (execParams.type) {
+    case TransactionType.TransferAsset: {
+      return algosdk.makeAssetTransferTxnWithSuggestedParams(
+        execParams.fromAccount.addr,
+        execParams.toAccountAddr,
+        execParams.payFlags.closeRemainderTo,
+        undefined,
+        execParams.amount,
+        note,
+        execParams.assetID,
+        suggestedParams);
+    }
+    case TransactionType.TransferAlgo: {
+      return algosdk.makePaymentTxnWithSuggestedParams(
+        execParams.fromAccount.addr,
+        execParams.toAccountAddr,
+        execParams.amountMicroAlgos,
+        execParams.payFlags.closeRemainderTo,
+        note,
+        suggestedParams);
+    }
+    case TransactionType.ClearSSC: {
+      return algosdk.makeApplicationClearStateTxn(
+        execParams.fromAccount.addr,
+        suggestedParams,
+        execParams.appId,
+        parseSSCAppArgs(execParams.appArgs),
+        execParams.accounts,
+        execParams.foreignApps,
+        execParams.foreignAssets,
+        note,
+        execParams.lease,
+        execParams.rekeyTo
+      );
+    }
+    case TransactionType.DeleteSSC: {
+      return algosdk.makeApplicationDeleteTxn(
+        execParams.fromAccount.addr,
+        suggestedParams,
+        execParams.appId,
+        parseSSCAppArgs(execParams.appArgs),
+        execParams.accounts,
+        execParams.foreignApps,
+        execParams.foreignAssets,
+        note,
+        execParams.lease,
+        execParams.rekeyTo
+      );
+    }
+    case TransactionType.CallNoOpSSC: {
+      return algosdk.makeApplicationNoOpTxn(
+        execParams.fromAccount.addr,
+        suggestedParams,
+        execParams.appId,
+        parseSSCAppArgs(execParams.appArgs),
+        execParams.accounts,
+        execParams.foreignApps,
+        execParams.foreignAssets,
+        note,
+        execParams.lease,
+        execParams.rekeyTo);
+    }
+    case TransactionType.CloseSSC: {
+      return algosdk.makeApplicationCloseOutTxn(
+        execParams.fromAccount.addr,
+        suggestedParams,
+        execParams.appId,
+        parseSSCAppArgs(execParams.appArgs),
+        execParams.accounts,
+        execParams.foreignApps,
+        execParams.foreignAssets,
+        note,
+        execParams.lease,
+        execParams.rekeyTo
+      );
+    }
+    default: {
+      throw new TealError(ERRORS.TRANSACTION.UNKNOWN_TRANSACTION_TYPE, { transaction: transactionType });
+    }
+  }
 }
