+++ packages/runtime/test/integration/deleteApp.ts
@@ -1,9 +1,9 @@
-import { SignType, SSCCallsParam, TransactionType } from "@algorand-builder/algob/build/types";
 import { assert } from "chai";
 
 import { ERRORS } from "../../src/errors/errors-list";
 import { Runtime, StoreAccount } from "../../src/index";
 import { ALGORAND_ACCOUNT_MIN_BALANCE } from "../../src/lib/constants";
+import { SignType, SSCCallsParam, TransactionType } from "../../src/types";
 import { expectTealError } from "../helpers/errors";
 import { getProgram } from "../helpers/files";
 import { useFixture } from "../helpers/integration";
