+++ packages/runtime/test/integration/stateful-counter.ts
@@ -1,9 +1,9 @@
-import { ExecParams, SignType, TransactionType } from "@algorand-builder/algob";
 import { assert } from "chai";
 
 import { Runtime, StoreAccount } from "../../src/index";
 import { BIGINT1 } from "../../src/interpreter/opcode-list";
 import { ALGORAND_ACCOUNT_MIN_BALANCE } from "../../src/lib/constants";
+import { ExecParams, SignType, TransactionType } from "../../src/types";
 import { getProgram } from "../helpers/files";
 import { useFixture } from "../helpers/integration";
